test: move demangle metadata for moved symbols to a separate file
